refactor: move types from networkMessages to own file
Types of UDP-messages and their tests should be in their own file.
